feat: format snake.cpp
