+++ snake/makefile
@@ -1,12 +1,12 @@
-CC = gcc
+CC = g++
 
 all: snake
 
 snake: linklist.o set_ticker.o snake.o res.o xc_queue.o
 	$(CC) -o snake -g linklist.o set_ticker.o snake.o res.o xc_queue.o -lcurses
 
-snake.o: snake.c snake.h
-	$(CC) -c -g snake.c
+snake.o: snake.cpp snake.h
+	$(CC) -c -g snake.cpp
 
 set_ticker.o: set_ticker.c
 	$(CC) -c -g set_ticker.c
